+++ toolchains/egret-webpack-bundler/src/loaders/theme.ts
@@ -2,6 +2,8 @@ import { EuiCompiler } from '@egret/eui-compiler';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as webpack from 'webpack';
+import { LineEmitter } from './inline-loader';
+import { AbstractInlinePlugin } from './inline-loader/AbstractInlinePlugin';
 import * as utils from './utils';
 
 type ThemePluginOptions = {
@@ -9,97 +11,57 @@ type ThemePluginOptions = {
     output: 'inline' | 'standalone'
 }[]
 
-export default class ThemePlugin {
+export default class ThemePlugin extends AbstractInlinePlugin {
     private options: Required<ThemePluginOptions>;
-    private isFirst = true;
 
     constructor() {
+        super();
         this.options = [{
             dirs: ['resource/eui_skins', 'resource/skins'],
             output: 'inline'
         }];
     }
 
-    private errors!: any[];
-    private compiler!: webpack.Compiler;
-
-    public apply(compiler: webpack.Compiler) {
-
-        this.errors = [];
-        this.compiler = compiler;
-        const dirs = this.options[0].dirs.map((dir) => path.join(compiler.context, dir));
-        const pluginName = this.constructor.name;
+    createLineEmitter(compiler: webpack.Compiler) {
         const euiCompiler = new EuiCompiler(compiler.context);
         const theme = euiCompiler.getThemes()[0];
-        const outputFilename = theme.filePath.replace('.thm.json', '.thm.js');
-        const thmJSPath = path.join(compiler.context, outputFilename);
-        utils.addWatchIgnore(compiler, thmJSPath);
+        const dirs = this.options[0].dirs.map((dir) => path.join(compiler.context, dir));
+        this.addFileDependency(path.join(compiler.context, theme.filePath));
+        for (let dir of dirs) {
+            this.addContextDependency(dir);
+        }
 
         const requires = theme.data.exmls.map((exml) => `require("./${path.relative(path.join(compiler.context, 'src'), exml).split('\\').join('/')}");`);
-        const themeJsContent = `window.skins = window.skins || {};
-    window.generateEUI = window.generateEUI || {
-      paths: {},
-      styles: undefined,
-      skins: ${JSON.stringify(theme.data.skins, null, '\t')},
-    };
-    ${requires.join('\n')}
-    module.exports = window.generateEUI;
-    `;
+        const themeJsContent = [
+            `window.skins = window.skins || {};`,
+            `window.generateEUI = window.generateEUI || {`,
+            `   paths: {},styles: undefined,`,
+            `   skins: ${JSON.stringify(theme.data.skins, null)},`,
+            `};`,
 
-        //   if (utils.isHot(this.compiler)) {
-        //     content += '\nif (module.hot) { module.hot.accept(); }';
-        //   }
-        const beforeRun = async (compilr: webpack.Compiler) => {
-
-            this.errors = [];
-            try {
-                const euiCompiler = new EuiCompiler(compiler.context, 'debug');
-                const result = euiCompiler.emit();
-
-                // eslint-disable-next-line global-require
-                const fs = require('fs');
-                const filename = path.join(this.compiler.context, result[0].filename);
-                const content = result[0].content;
-
-                fs.writeFileSync(filename, content, 'utf-8');
-                // 更新文件系统缓存状态
-
-                if (this.isFirst) {
-                    const inlineLoaderRule: webpack.RuleSetRule = {
-                        test: /Main\.ts/,
-                        include: path.join(compilr.context!, 'src'),
-                        loader: require.resolve('./inline-loader'),
-                        options: { content: themeJsContent }
-                    };
-
-                    compiler.options.module?.rules.push(inlineLoaderRule);
-                    this.isFirst = false;
-                }
+        ].concat(requires).concat([
+            `module.exports = window.generateEUI;`
+        ]);
+        return {
+            emitLines: () => {
+                return themeJsContent
             }
-            catch (error) {
-                // // 写入错误信息
-                this.errors.push(error);
-            }
+        } as LineEmitter
+    }
 
-        };
+    onChange(compilation: webpack.Compilation) {
+        const compiler = compilation.compiler;
+        const euiCompiler = new EuiCompiler(compiler.context, 'debug');
 
-        compiler.hooks.thisCompilation.tap(pluginName, (compilation: webpack.Compilation) => {
-            if (this.errors.length) {
-                compilation.errors.push(...this.errors);
-            }
-        });
+        const theme = euiCompiler.getThemes()[0];
+        const themeFile = path.join(compiler.context, theme.filePath)
+        if (utils.fileChanged(compiler, themeFile)) {
+            const result = euiCompiler.emit();
+            const filename = path.join(compiler.context, result[0].filename);
+            const content = result[0].content;
+            fs.writeFileSync(filename, content, 'utf-8');
+        }
 
-        compiler.hooks.watchRun.tapPromise(pluginName, beforeRun);
-        compiler.hooks.beforeRun.tapPromise(pluginName, beforeRun);
-
-        // 监听文件目录
-        compiler.hooks.afterCompile.tap(pluginName, (compilation) => {
-            dirs.forEach((item) => {
-                if (fs.existsSync(item)) {
-                    compilation.contextDependencies.add(item);
-                }
-            });
-        });
     }
 }