+++ toolchains/egret-webpack-bundler/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @egret/egret-webpack-bundler
 
-This log was last generated on Thu, 28 Jan 2021 04:53:08 GMT and should not be manually modified.
+This log was last generated on Fri, 29 Jan 2021 06:03:57 GMT and should not be manually modified.
+
+## 2.0.3
+Fri, 29 Jan 2021 06:03:57 GMT
+
+### Patches
+
+- 改善 exml-loader 和 typescript-legacy 模式的增量编译触发逻辑
 
 ## 2.0.2
 Thu, 28 Jan 2021 04:53:08 GMT